requirements.txt
# Base Django requirement
Django>=4.0,<5.0  # Or your preferred Django version range

# Required for ImageField
Pillow>=9.0,<11.0 # Check compatibility with your Python/Django version

# Additional requirements
requests>=2.28.0  # For handling HTTP requests
openpyxl>=3.0.0  # For Excel export
reportlab>=3.6.0  # For PDF export
django-import-export>=2.5.0  # For admin export functionality 